feat: make generic.is_url() accept dollar-quoted userinfo fragments
Part of supporting authenticated remote styles requires a way to specify
authinfo without necessarily hardcoding a sensitive token in the config
file.

This is done by interpolating an environment variable if the url
username fragment begins with a `$`, but that fails the original
`nitpick.generic.URL_RE` pattern, because `\w` weirdly doesn't match
`$`.